# OPENAPI_MVP_V1.yaml
openapi: 3.0.3
info:
  title: Kinetica MVP API
  version: 1.0.0
  description: >-
    MVP contract for web/mobile clients and AI service parallel development.
servers:
  - url: http://localhost:8080
paths:
...
  /translations:
    get:
      tags: [Translations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List translation requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslationResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      tags: [Translations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTranslationRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /translations/{id}:
    get:
      tags: [Translations]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Translation status/result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    put:
      tags: [Translations]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTranslationRequest'
      responses:
        '200':
          description: Translation updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '400': { $ref: '#/components/responses/BadRequestError' }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Translations]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequestError:
      description: Bad request (validation or business rule error).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    UnauthorizedError:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    ForbiddenError:
      description: Authenticated but without enough privileges.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    NotFoundError:
      description: Resource was not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorResponse' }
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message, details, timestamp]
      properties:
        code: { type: string, example: BAD_REQUEST }
        message: { type: string, example: Request could not be processed. }
        details:
          type: array
          items: { type: string }
        timestamp: { type: string, format: date-time }
...
    CreateTranslationRequest:
      type: object
      required: [direction, sourceText]
      properties:
        direction:
          type: string
          enum: [SIGN_TO_TEXT, TEXT_TO_SIGN]
        sourceText:
          type: string
    UpdateTranslationRequest:
      type: object
      required: [sourceText]
      properties:
        sourceText: { type: string }
    TranslationResponse:
      type: object
      properties:
        requestId: { type: integer, format: int64 }
        status:
          type: string
          enum: [PENDING, DONE, FAILED]
        direction:
          type: string
          enum: [SIGN_TO_TEXT, TEXT_TO_SIGN]
        textOutput: { type: string, nullable: true }
        glossOutput:
          type: string
          nullable: true
          description: Intermediate LSP gloss used in the translation pipeline.
        signOutputRef:
          type: string
          nullable: true
          description: >-
            Reference to sign representation for frontend rendering.
            For 3D avatar flows, this can point to sign-sequence JSON such as
            sign://catalog/basic-sequence or an asset key to fetch animation instructions.
        confidence: { type: number, format: double, nullable: true }
        warning: { type: string, nullable: true, example: LOW_CONFIDENCE }
        modelVersion: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...